[Looker] Adds LookML autogeneration script for all observation views (Recidiviz/recidiviz-data#40091)
## Description of the change

Adds an autogeneration script/package for generating LookML views for
all observations, mirroring the same structure that exists in BQ. I use
the output of this script to enable building explores off of any
observations (see looker PR:
https://github.com/Recidiviz/looker/pull/726) which unblocks a bug fix
to some broken tiles in the Insights readiness dash.

We'll want to eventually build out some sort of looker native view that
mimics the functionality of all_person_events, which is being referenced
by `unified_spans` -- this is a precursor to that.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 21d15b3ca28fbd5c334754c3a849138d039bbdcb