[US_UT] Map new assessment type (Recidiviz/recidiviz-data#40412)
## Description of the change

In the raw data we imported yesterday, there was a new type of
assessment. Since it wasn't mapped in the ingest mapping, it is causing
ingest to crash. This maps the new test name to `INTERNAL_UNKNOWN` since
it is not one that we currently need to reference or understand.

The run mappings script succeeds after this change - I am running an
ingest pipeline now to make sure there aren't other errors hiding behind
this one

https://console.cloud.google.com/dataflow/jobs/us-east1/2025-04-02_09_49_47-503031784143362170;graphView=1?project=recidiviz-staging&invt=AbtsdA&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22))

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d1fed0cc61a16e9e5d1f77fcd09660c2946b6856