[US_UT] Clear raw data import errors  (Recidiviz/recidiviz-data#40433)
## Description of the change

There are many raw data import errors in UT staging today. These have
two causes that are handled here:
- Files with <10 rows have one row more/less than they have previously,
which exceeds the 10% threshold allowed. This PR bumps the stable counts
threshold to 50% for those files specifically.
- Every column in the `sent_adj_rsn_cd` was entirely NULL in the most
recent transfer. Since that is not helpful and we do not use this file
anyway, this PR deprecates it. will follow steps in [this
doc](https://www.notion.so/recidiviz/Managing-raw-data-IMPORT-03eac842214f44dda0a4bd518c11f043)
to fully deprecate!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 790b063dbc1562e0649d2bebca919d7add15ddb9