[Insights] Add helper for most recent attrs cte (Recidiviz/recidiviz-data#40137)
## Description of the change

Adds a util fn for the most recent officer attrs CTE.

This CTE is currently used in two insights views and we'll additionally
start consuming it in the `supervision_officer_metrics` view when we
consolidate the officer exclusion logic to rows in the metrics view (see
PR [here](https://github.com/Recidiviz/recidiviz-data/pull/40158)).

This is just a refactor PR to make Recidiviz/recidiviz-data#40158 simpler.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues
 Recidiviz/recidiviz-data#38094

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 93c5498b67ca90bc7147643edf7cc03ebed8f7c9