[US_PA][Workflows] Display pending charge information on sidebar (Recidiviz/recidiviz-data#40340)
## Description of the change

This PR adds pending charges to the sidebar of both admin supervision
and special circumstances supervision opportunities.

For admin supervision, we want to flag all pending charges since by
[policy](https://docs.google.com/document/d/1pEZx9SkWFQT2B1f-D38MQhU8lkaDXicONAprV2r4fz0/edit?tab=t.0),
reentrants "charged with a misdemeanor or felony and not detained
pending the disposition of the new criminal charges" are not eligible
for admin supervision. For special circ supervision, we only want to
flag pending violent offenses for reentrants that are currently
considered to be serving non-violent cases. We're flagging these in the
sidebar instead of writing them into eligibility criteria for reasons
I've detailed more
[here](https://www.notion.so/recidiviz/Are-pending-charges-even-real-1ae7889f4d1980fa85ded635b130036f)
if you're curious, but basically I'm not super confident in the data
quality and our TT suggested flagging them on the sidebar instead to
"allows agents and supervisors to apply discretion, verify the nature of
the charge, and determine whether it should impact eligibility based on
the broader context of the case."

I also had to move the admin supervision offense logic into a helper
function which is why there's like a million lines of code changes but
ignore those :) the actual changes are in the case notes helpers at the
bottom of `us_pa_query_fragments`.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#38995

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e2652bc15c07a329cbfe988057fea96a6aa987f6