[US_UT] Address more field changes in raw data (Recidiviz/recidiviz-data#40518)
## Description of the change

I spaced out yesterday and did not test that import would succeed for
every previously failing file before merging
https://github.com/Recidiviz/recidiviz-data/pull/40433. This PR
addresses the few issues that remained after that first set of issues
was cleared.

This PR also marks the `applc_usr` table as incremental, even though I
don't believe it is - this is to circumvent an
[issue](https://github.com/Recidiviz/recidiviz-data/issues/40396) where
one user who has been assigned a supervision caseload in the past no
longer appears in the most recent version of the table. This is causing
ingest to fail when it ingests staff only from the most recently
transferred version of the table. If we mark the file as incremental, it
will include the most recent version of every PK in the table
(`usr_id`), which will resolve this issue.
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ebd4ad6efef68b3a18829ff460aa0c434af4e75a