[Insights] Add include_in_outcomes metric rows to SOM view (Recidiviz/recidiviz-data#40494)
## Description of the change

Starts calculating a `include_in_outcomes` metric row for every insights
supervision officer.

This PR is part 1 of 2 to start reading the officer entity's
`include_in_outcomes` value off of metric rows, rather than a
`include_in_outcomes` column. The original PR to do this was reverted
because we didn't separate the BE change from the view change - and the
BE was not backwards compatible. That PR has a very in depth description
of the reasoning behind this change: Recidiviz/recidiviz-data#40158.

This PR does not touch the querier and only contains the view
modifications that do not impact our BE queries, setting us up to start
querying for metric rows instead of columns in the next PR: Recidiviz/recidiviz-data#40495 .
That PR should not land until this goes to prod.

Sandbox is at `gina2`, confirmed that we're getting one new
`include_in_outcomes` metric row per officer and the existing
`include_in_outcomes` column is still populated on outcomes metrics.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#38094

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1b30e58d23fc4a088fa80f450a26357bff3462fd