[US_IX][Workflows] Add `eligible_date` field to Idaho opp records (Recidiviz/recidiviz-data#40188)
## Description of the change

Add an `eligible_date` column to Idaho opportunity records for non-CRC
opportunities.

Background: We would like to start including an "eligibility date" (not
necessarily the eligibility date per se, but some relevant date) on
opportunity records, in preparation for launching Table View in the
Workflows tool. In Table View, we will both display the date and enable
the user to sort by this column. For more info, see the ticket.

All of the affected opportunities already have an "eligibility date"
defined in code in the frontend, so this PR will lead to no changes in
what is displayed - but it will allow us to later remove the frontend in
case we want to start displaying a different date, and having the date
in a standarduzed place means that new opportunities being developed can
have this date start showing up "for free".

After loading to sandbox with this command, there appears to be no
change in the number of opportunity records surfaced for any of these
opportunities:
```
python -m recidiviz.tools.load_views_to_sandbox \
       --sandbox_dataset_prefix ryan_id_eligible_date \
       --state_code_filter US_IX auto \
       --load_changed_views_only
```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#39379

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 12df20606ae361fa1d2e7ec1a9ff2fe0c2ca103d