[Insights] Add rate_denominator to OutliersMetricConfig and view helper fn (Recidiviz/recidiviz-data#40503)
## Description of the change

Adds a config field for `rate_denominator` to the OutliersMetricConfig
and consumes the field in the SOM view.

When calculating metric values, for `RATE` rows, we historically have
used the `avg_daily_population` field as the denominator to calculate
the rate. However, as we work towards splitting MI's metrics up between
parole v probation, we need the option to have different denominators
depending on the metric. E.g. for parole metrics, we'd provide
`avg_population_parole` as the metric rate denominator. The default
should still be `avg_daily_population`, since that's what almost all the
metrics will continue to use.

Next up we'll add a FV for splitting the parole/probation metrics and
use that to handle any necessary BE logic changes before adding MI
outlier metric configs for the new metrics.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#39162

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8cfc70da845b78dfc3b885f396352bd5b7feedd3