[Admin Panel] Fix custom permissions highlighting (Recidiviz/recidiviz-data#40165)
## Description of the change

Highlighting of custom permissions broke when we implemented multiple
roles because the users' merged permissions were still being compared to
the default permissions for a single role rather than the merged
permissions. This PR fixes that issue by generating what the merged
default permissions should look like (essentially implementing [the
merge_permissions
function](https://github.com/Recidiviz/recidiviz-data/blob/c6690cf6ef20da479f46b048de3faab11fd2adb3/recidiviz/auth/helpers.py#L128-L159)
in the frontend) and comparing that to the users' permissions. I assumed
that if a particular route/fv is not included in either of the
permissions objects being compared, that is equivalent to that route/fv
being set to 'false' in the other permissions object. I also added some
tests since there is a lot going on there.

The other change made is that the only the permission that is different
from the default permissions is highlighted instead of the entire cell.
This does mean that if a custom permission is removing access instead of
adding it, it won't show up--is that a problem? I'm guessing it's more
common for the custom permissions to be adding permissions to the
default rather than removing them.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-dashboardsRecidiviz/recidiviz-data#7630

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 695b0b67c7f4e9e3275ba7522bc17ad7b1f03d21