[JII] Add group_name and workflow_execution_id to new docs to help migration (Recidiviz/recidiviz-data#40567)
## Description of the change

Adds `group_name` to jii (person) level documents and
`workflow_execution_id` to message level documents in preparation for
the next manual launch on 04/08/25.

The addition of these fields is simply to make the migration of these
documents to the new JII CloudSql db easier. Rather than migrating all
(absolute) documents in the Firestore database again (we completed this
last week), we will only migrate the new or changed (delta) documents
that include these fields. When it comes time to migrate these
documents, we will be able to filter by these new fields.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

Co-authored-by: Michelle Orden <michelleorden@Michelles-MacBook-Pro.local>
GitOrigin-RevId: ece35c5f98c6adf5aaccd353b4f33e2a5acab104